requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
httpx==0.27.2
tenacity==9.0.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
